Delete custom security config for now
It is very likely that we won't use JWT here, as we won't use the
Tobira uploader, Studio or Editor integration. And then I rather have
fewer custom config files to keep Opencast as vanilla as possible.

If we ever want to add it again, we can just revert this commit.