Upgrade Opencast to 18.1
Nothing really noteworthy about the update. I updated Java to 21 on
the server manually. I also updated whisper.cpp so that the binary
`whisper-cli` is available, so that we can use more of the OC default
config.

The encoding profile file became unnecessary already a while ago, since
OC does the audio encoding for whisper internally.